requirements.txt
asgiref==3.3.4
coverage==5.5
Django==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
flake8==3.9.2
flake8-isort==4.0.0
isort==5.8.0
mccabe==0.6.1
Pillow==8.2.0
pkg-resources==0.0.0
pycodestyle==2.7.0
pydub==0.25.1
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.1
testfixtures==6.17.1
